requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
pandas>=1.3.0
scikit-learn>=1.0.0
networkx>=2.6.0
mne>=1.0.0
nilearn>=0.8.0
pymc>=4.0.0
torch>=1.10.0
statsmodels>=0.13.0
joblib>=1.1.0
tqdm>=4.62.0
pyyaml>=6.0
h5py>=3.3.0